fixup! perf(core): detect existing signal dependency without checking all producer links
